requirements.txt
streamlit==1.13.0
scikit-learn==1.0.2
librosa==0.9.2
seaborn==0.11.2
matplotlib==3.5.2
umap-learn==0.5.3